requirements.txt
astroid==2.4.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.7
django-filter==2.2.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
isort==4.3.21
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.1
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.3
Pillow==7.1.2
psycopg2==2.8.5
pylint==2.5.0
pyparsing==2.4.7
pyTelegramBotAPI==3.7.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
response==0.4.0
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
static3==0.7.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
